move operator supervisor into operator app
we add the inotify application to the operator supervisor so they can be started
in the correct order inotify -> operator, and we set the policy to rest for one
since if inotify crashes we need to restart the operator so it can connect to
it again.
update the *.app.src, start_link/0,1 in operator to reflect the above.